fixing issue on branch service

+++ src/main/java/com/goDelivery/goDelivery/dto/branch/BranchCreationDTO.java
@@ -66,7 +66,7 @@ public class BranchCreationDTO {
     
     // Delivery settings
     private Boolean deliveryAvailable = false;
-    private Float deliveryRadius; // in km
+    private Double deliveryRadius; // in km
     private Float minimumOrderAmount;
     private Float deliveryFee;
